Refactor: change method names, move methods into more suitable place.
